reuse CCTX type
